+++ src/main/java/com/letsTravel/LetsTravel/domain/community/entity/Board.java
@@ -1,4 +1,4 @@
-package com.letsTravel.LetsTravel.domain.community;
+package com.letsTravel.LetsTravel.domain.community.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
