+++ src/main/java/com/letsTravel/LetsTravel/domain/community/entity/Comment.java
@@ -1,6 +1,6 @@
-package com.letsTravel.LetsTravel.domain.community;
+package com.letsTravel.LetsTravel.domain.community.entity;
 
-import com.letsTravel.LetsTravel.domain.profile.Profile;
+import com.letsTravel.LetsTravel.domain.profile.entity.Profile;
 import com.letsTravel.LetsTravel.domain.util.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
