+++ src/main/java/com/letsTravel/LetsTravel/domain/map/entity/Place.java
@@ -1,9 +1,7 @@
-package com.letsTravel.LetsTravel.domain.map;
+package com.letsTravel.LetsTravel.domain.map.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.math.BigDecimal;
 
 @Builder
 @Entity
