+++ src/main/java/com/letsTravel/LetsTravel/domain/map/entity/PlaceTranslated.java
@@ -1,4 +1,4 @@
-package com.letsTravel.LetsTravel.domain.map;
+package com.letsTravel.LetsTravel.domain.map.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,21 +8,22 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class LocalityTranslated {
+public class PlaceTranslated {
 
     @Id
-    @Column(columnDefinition = "MEDIUMINT")
+    @Column(columnDefinition = "INT")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
-    @Column(columnDefinition = "VARCHAR(58)", nullable = false)
-    private String localityName;
+    @Column(columnDefinition = "VARCHAR(85)", nullable = false)
+    private String placeName;
 
     @Column(columnDefinition = "VARCHAR(2)", nullable = false)
     private String languageCode;
 
     @ManyToOne
-    @JoinColumn(name = "locality_id", nullable = false)
-    private Locality locality;
+    @JoinColumn(name = "place_id", nullable = false)
+    private Place place;
 
 }
+
