--- src/main/java/com/letsTravel/LetsTravel/domain/profile/UserRole.java
@@ -1,6 +1,0 @@
-package com.letsTravel.LetsTravel.domain.profile;
-
-public enum UserRole {
-    ROLE_USER, ROLE_ADMIN;
-
-}
