+++ src/main/java/com/letsTravel/LetsTravel/domain/profile/entity/Profile.java
@@ -1,4 +1,4 @@
-package com.letsTravel.LetsTravel.domain.profile;
+package com.letsTravel.LetsTravel.domain.profile.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
