+++ src/main/java/com/letsTravel/LetsTravel/domain/profile/entity/UserRole.java
@@ -1,0 +1,6 @@
+package com.letsTravel.LetsTravel.domain.profile.entity;
+
+public enum UserRole {
+    ROLE_USER, ROLE_ADMIN;
+
+}
